requirements.txt
pandas>=2.0.0
numpy>=1.24.0
dask[dataframe]>=2023.1.0
pyyaml>=6.0
python-dateutil>=2.8.2
tqdm>=4.65.0
ydata-profiling>=4.0.0
openpyxl>=3.0.0
scikit-learn>=1.2.0
fsspec>=2023.1.0
pyarrow>=12.0.0
loguru>=0.6.0
pytest>=7.2.0